requirements.txt
pydantic==1.9.0
jsonpath==0.82
uvicorn==0.17.4
PyMySQL==1.0.2
allure_python_commons==2.8.0
colorlog==5.0.1
pytest==6.2.3
requests==2.21.0
fastapi==0.78.0
PyYAML==6.0
